requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pydantic==1.10.12
google-generativeai==0.14.0
requests==2.32.3
motor==3.1.1
streamlit==1.38.0  # optional - remove if not using
